+++ src/gui/about.cpp
@@ -8,20 +8,6 @@
 
 #include "gui/imguiext.hpp"
 
-void drawLink(const char* label, const char* link) {
-    ImGui::PushID(label);
-    bool copy = ImGui::SmallButton("Copy");
-    ImGui::PopID();
-
-    ImGui::SameLine();
-    ImGui::Text("%s:", label);
-    ImGui::SameLine();
-
-    if (copy) ImGui::LogToClipboard();
-    ImGui::Text("https://%s", link);
-    if (copy) ImGui::LogFinish();
-}
-
 void drawAboutWindow(bool& open) {
     if (!open) return;
 
@@ -60,14 +46,16 @@ void drawLinksWindow(bool& open) {
     if (!open) return;
 
     using namespace ImGuiExt::Literals;
-    ImGui::SetNextWindowSize(40_fh * 10_fh, ImGuiCond_FirstUseEver);
+    ImGui::SetNextWindowSize(20_fh * 10_fh, ImGuiCond_FirstUseEver);
     if (!ImGui::Begin("Links", &open)) {
         ImGui::End();
         return;
     }
 
-    drawLink("Repository", "github.com/WhaleConnect/whaleconnect");
-    drawLink("Changelog", "github.com/WhaleConnect/whaleconnect/blob/main/docs/changelog.md");
+    ImGui::TextWrapped("These are helpful links to get information and support.");
+    ImGui::TextLinkOpenURL("Repository", "https://github.com/WhaleConnect/whaleconnect");
+    ImGui::SameLine();
+    ImGui::TextLinkOpenURL("Changelog", "https://github.com/WhaleConnect/whaleconnect/blob/main/docs/changelog.md");
 
     ImGui::End();
 }